+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.54.0 - 2025-05-13
+
+#### Enhancements
+- Added new off-market publishers for Eurex, and European Energy Exchange (EEX)
+- Increased live subscription symbol chunking size
+- Upgraded `databento-dbn` to 0.34.0
+
 ## 0.53.0 - 2025-04-29
 
 #### Enhancements